feat(better_deployments): configs
